+++ emulation_system/emulation_system/compose_file_creator/input/hardware_models/hardware_specific_attributes.py
