+++ emulation_system/emulation_system/compose_file_creator/input/hardware_models/modules/heater_shaker_module.py
@@ -1,0 +1,24 @@
+from typing_extensions import Literal
+
+from pydantic import Field
+
+from emulation_system.compose_file_creator.input.hardware_models\
+    .hardware_specific_attributes import HardwareSpecificAttributes
+from emulation_system.compose_file_creator.input.hardware_models\
+    .modules.module_model import ModuleModel
+
+from emulation_system.compose_file_creator.input.settings import (
+    HeaterShakerModes
+)
+
+
+class HeaterShakerModuleAttributes(HardwareSpecificAttributes):
+    mode: HeaterShakerModes = HeaterShakerModes.SOCKET
+
+
+class HeaterShakerModuleModel(ModuleModel):
+    hardware: Literal["heater-shaker"]
+    hardware_specific_attributes: HeaterShakerModuleAttributes = Field(
+        alias="hardware-specific-attributes",
+        default=HeaterShakerModuleAttributes()
+    )
