+++ emulation_system/emulation_system/compose_file_creator/input/hardware_models/modules/module_model.py
@@ -1,0 +1,6 @@
+from emulation_system.compose_file_creator.input.hardware_models\
+    .hardware_model import HardwareModel
+
+
+class ModuleModel(HardwareModel):
+    pass
