--- emulation_system/emulation_system/compose_file_creator/input/models/hardware_specific_attributes/__init__.py
@@ -1,15 +1,0 @@
-from emulation_system.compose_file_creator.input.models.hardware_specific_attributes.\
-    ot2 import OT2Attributes
-from emulation_system.compose_file_creator.input.models.hardware_specific_attributes.\
-    heater_shaker_module import HeaterShakerModuleAttributes
-from emulation_system.compose_file_creator.input.models.hardware_specific_attributes.\
-    temperature_module import TemperatureModuleAttributes
-from emulation_system.compose_file_creator.input.models.hardware_specific_attributes.\
-    thermocycler_module import ThermocyclerModuleAttributes
-
-__all__ = [
-    "OT2Attributes",
-    "HeaterShakerModuleAttributes",
-    "ThermocyclerModuleAttributes",
-    "TemperatureModuleAttributes"
-]
